Kernel: Rewrite ProcFS.
Now the filesystem is generated on-the-fly instead of manually adding and
removing inodes as processes spawn and die.

The code is convoluted and bloated as I wrote it while sleepless. However,
it's still vastly better than the old ProcFS, so I'm committing it.

I also added /proc/PID/fd/N symlinks for each of a process's open fd's.

+++ LibC/sys/types.h
@@ -15,7 +15,7 @@ typedef __intmax_t intmax_t;
 
 typedef uint32_t uid_t;
 typedef uint32_t gid_t;
-typedef int pid_t;
+typedef int16_t pid_t;
 
 typedef uint32_t size_t;
 typedef int32_t ssize_t;
